adding eval_fiber_model method to Image class & implementation across
